Add custom sources and layers
